Improve randomizer logic
A list of 0 to num_of_players is generated when the session is created.
Each time a player joins, a random value is taken from the list and that value is then removed from the list.
The random value is then used to index into the num_of_players to assign a role and name